Initial probes are trained
On every frame all nexuses are checked for nexuses that completed and
are not training probes. For every such nexus, a probe is trained if
there is need and resources. A need of probes happens when there is a
mineral patch around the base with less than 2 probes on it.